+++ plugins/ldap/src/plugin.ts
@@ -1,5 +1,5 @@
 // imports here
-import { AppSh, AppShPlugin } from "app-sh";
+import { AppSh, AppShPlugin, AppShConfig } from "app-sh";
 
 import ldap from "ldapjs";
 
@@ -8,6 +8,7 @@ import ldap from "ldapjs";
 // Config consts here
 export type LdapConfig = {
   appSh?: AppSh;
+  appShConfig?: AppShConfig;
 
   adServer: string;
   userDnBase: string;
@@ -35,6 +36,7 @@ export class Ldap extends AppShPlugin {
     super({
       name: "ldap",
       appSh: config.appSh,
+      appShConfig: config.appShConfig,
       // NOTE: PLUGIN_VERSION is replaced with package.json#version by a
       // rollup plugin at build time
       pluginVersion: "PLUGIN_VERSION",