+++ plugins/ldap/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @app-sh/ldap
 
+## 1.2.0
+
+### Minor Changes
+
+- Added the ability to provide an AppSh config to the plugins
+
 ## 1.1.0
 
 ### Minor Changes